+++ checkout-ui-custom/src/_js/_v.custom.checkout.ui.js
@@ -356,15 +356,19 @@ class checkoutCustom {
 
   condensedTaxes(orderForm) {
     let customtax = orderForm.totalizers.filter(val => val.id=="CustomTax");
-    if(customtax.length<2) return false;
+    if(customtax && customtax.length<2) return false;
 
     let tooltip = `
       <div class="vcustom-customTax-resume">
        ${customtax.map(i => `<p class="vcustom-customTax-resume__i"><span class="n">${i.name}</span><span class="v">${orderForm.storePreferencesData.currencySymbol} ${(i.value/100).toFixed(2)}</span></p>`).join("")}
       </div>
     `;
-    $("tr.CustomTax.CustomTax--total").find(".vcustom-customTax-tot").remove();
-    $("tr.CustomTax.CustomTax--total").find(".info").append(`<div class="vcustom-customTax-tot"><span>?</span> ${tooltip}</div>`)
+
+    let customTaxElem = $("tr.CustomTax.CustomTax--total");
+    if(customTaxElem.length) {
+      customTaxElem.find(".vcustom-customTax-tot").remove();
+      customTaxElem.find(".info").append(`<div class="vcustom-customTax-tot"><span>?</span> ${tooltip}</div>`)
+    }
   }
 
   update(orderForm) {