+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Tooltip displaying custom taxes
+
 ## [0.0.23] - 2020-10-22
 
 ### Added